Work around limitations of Python 2 distutils and AX_PYTHON_DEVEL
Python 2 doesn't have the LDVERSION, so AX_PYTHON_DEVEL has to form
PYTHON_LIBS using plain VERSION, which doesn't have the _d suffix.
There doesn't seem to be a good way to find out that we need this,
so I can't upstream a fix to autoconf-archive. However, the -dbg
flavour is mostly a Debianism anyway, and in the long term the
problem will be fixed by Python 2 going away.

Signed-off-by: Simon McVittie <smcv@collabora.com>