Fix issue with service's methods without return which called via D-Bus
If declared method has no out_signature and does not return any value, sending the reply via d-bus cause errors. To resolve this problem a new parameter introduced in the 'method' decorator: 'return_none', with default value of 'False'. If this parameter is set to 'True', empty reply will be sent as it was before.